# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Pivot API",
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/elections/{electionId}/candidates": {
            "get": {
                "tags": [
                    "Candidates"
                ],
                "summary": "View candidates for an election",
                "operationId": "candidateIndex",
                "parameters": [
                    {
                        "name": "electionId",
                        "in": "path",
                        "description": "Election to get",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Candidate"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Candidates"
                ],
                "summary": "Add a candidate",
                "operationId": "createCandidate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "electionId",
                        "in": "path",
                        "description": "Election ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "payload",
                        "in": "body",
                        "description": "Candidate to add",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateCandidate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Candidate"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/api/elections": {
            "get": {
                "tags": [
                    "Election"
                ],
                "summary": "View all elections",
                "operationId": "electionIndex",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Election"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/elections/{electionId}": {
            "get": {
                "tags": [
                    "Election"
                ],
                "summary": "View information about an election",
                "operationId": "getElectionById",
                "parameters": [
                    {
                        "name": "electionId",
                        "in": "path",
                        "description": "Election to get",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ElectionWithCreator"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Candidate": {
            "properties": {
                "id": {
                    "description": "",
                    "type": "string",
                    "example": "123"
                },
                "election_id": {
                    "description": "",
                    "type": "string",
                    "example": "123"
                },
                "name": {
                    "description": "",
                    "type": "string",
                    "example": "Funds for road construction"
                },
                "created_at": {
                    "description": "",
                    "type": "string",
                    "example": "2017-01-01 11:00:00"
                },
                "updated_at": {
                    "description": "",
                    "type": "string",
                    "example": "2017-01-01 11:00:00"
                }
            },
            "type": "object",
            "xml": {
                "name": "Candidate"
            }
        },
...
        "CreateCandidate": {
            "properties": {
                "name": {
                    "description": "",
                    "type": "string",
                    "example": "Highland Avenue"
                }
            },
            "type": "object",
            "xml": {
                "name": "CreateCandidate"
            }
        },
        "Election": {
            "properties": {
                "id": {
                    "description": "",
                    "type": "string",
                    "example": "123"
                },
                "creator_id": {
                    "description": "",
                    "type": "string",
                    "example": "123"
                },
                "name": {
                    "description": "",
                    "type": "string",
                    "example": "Funds for road construction"
                },
                "created_at": {
                    "description": "",
                    "type": "string",
                    "example": "2017-01-01 11:00:00"
                },
                "updated_at": {
                    "description": "",
                    "type": "string",
                    "example": "2017-01-01 11:00:00"
                },
                "deleted_at": {
                    "description": "",
                    "type": "string",
                    "example": "2017-01-01 11:00:00"
                }
            },
            "type": "object",
            "xml": {
                "name": "Election"
            }
        },
        "ElectionWithCreator": {
            "properties": {
                "id": {
                    "description": "",
                    "type": "string",
                    "example": "123"
                },
                "creator_id": {
                    "description": "",
                    "type": "string",
                    "example": "123"
                },
                "name": {
                    "description": "",
                    "type": "string",
                    "example": "Funds for road construction"
                },
                "created_at": {
                    "description": "",
                    "type": "string",
                    "example": "2017-01-01 11:00:00"
                },
                "updated_at": {
                    "description": "",
                    "type": "string",
                    "example": "2017-01-01 11:00:00"
                },
                "deleted_at": {
                    "description": "",
                    "type": "string",
                    "example": "2017-01-01 11:00:00"
                },
                "creator": {
                    "description": "",
                    "$ref": "#/definitions/User"
                }
            },
            "type": "object",
            "xml": {
                "name": "ElectionWithCreator"
            }
        },
...
        "User": {
            "properties": {
                "id": {
                    "description": "",
                    "type": "string",
                    "example": "123"
                },
                "name": {
                    "description": "",
                    "type": "string",
                    "example": "John Doe"
                },
                "email": {
                    "description": "",
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "created_at": {
                    "description": "",
                    "type": "string",
                    "example": "2017-01-01 11:00:00"
                },
                "updated_at": {
                    "description": "",
                    "type": "string",
                    "example": "2017-01-01 11:00:00"
                }
            },
            "type": "object",
            "xml": {
                "name": "User"
            }
        }
    }
}